backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic
pydantic-settings
python-dotenv
ibm-watson
ibm-watsonx-ai
ibm-watsonx-orchestrate
pytest
pytest-asyncio
httpx
